feat: implement cardlist for data
